move systemClassPath method to JvmUtils
